[Improvement] Improve the reset factory script

	modified:   usr/local/zevenet/app/zbin/reset_factory.sh

+++ usr/local/zevenet/app/zbin/reset_factory.sh
@@ -2,6 +2,10 @@
 
 if_mgmt=$1
 
+CONF_DIR="/usr/local/zevenet/config";
+TEMPLATE_DIR="/usr/local/zevenet/share";
+
+
 show_usage() {
   echo usage: `basename $0`  interface_name
   exit 1
@@ -13,60 +17,80 @@ fi
 
 if [ ! -d "/sys/class/net/${if_mgmt}" ]; then
         echo "Specified interface does not exist"
-fi 
-echo "Parando procesos"
+fi
+echo "Stopping processes"
 for AP in $(ps aux | grep zen | grep -v grep | awk '{print $2}')
 do
         echo "Parando el proceso  $AP"
-        ps -ef | grep $AP |grep -v grep 
+        ps -ef | grep $AP |grep -v grep
         pkill $AP
 done
 kill -9 `ps aux | grep sec |grep -v grep | awk '{print $2}'`
 
-echo "Parando proceso cron y cherokee"
+echo "Stopping cron process"
 /etc/init.d/cron stop
+echo "Stopping cherokee process"
 /etc/init.d/cherokee stop
+echo "Stopping zevenet process"
 /etc/init.d/zevenet stop
 
-echo "Borrando ficheros de configuracion"
+echo "Deleting configuration files"
 rm -fr /var/log/*
 rm -fr /opt/*
 rm -fr /tmp/*
 rm -fr /usr/local/zevenet/logs/*
-rm -fr /usr/local/zevenet/backups/*
 rm -fr /usr/local/zevenet/www/zlbcertfile.pem
 rm -fr /usr/local/zevenet/app/zenrrd/rrd/*
 rm -fr /usr/local/zevenet/www/img/graphs/*
-rm -fr /usr/local/zevenet/config/if*
-rm -fr /usr/local/zevenet/config/notifications/*
-rm -fr /usr/local/zevenet/config/float.conf
-rm -fr /usr/local/zevenet/config/*.html
-rm -fr /usr/local/zevenet/config/*.cfg
-rm -fr /etc/apt/sources.list
 
-#borrar todo excpeto: zencert*, zlb-*, global.con, if_eth0_conf
+read -p "Do you want delete the backups [N|y]? " -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    rm -fr /usr/local/zevenet/backups/*
+fi
 
-echo "Limpiando apt"
-apt-get update
-apt-get clean
+mv ${CONF_DIR}/zencert* /tmp/
+rm -fr ${CONF_DIR}
+mv /tmp/zencert* $CONF_DIR
 
-echo "Borrando config"
-DGW=`grep defaultgw= /usr/local/zevenet/config/global.conf`
-sed -i "s/$DGW/\$defaultgw=\"\";/g" /usr/local/zevenet/config/global.conf
-DGW=`grep defaultgwif= /usr/local/zevenet/config/global.conf`
-sed -i "s/$DGW/\$defaultgwif=\"\";/g" /usr/local/zevenet/config/global.conf
+#borrar todo excpeto: zencert*, zlb-*, global.conf, if_eth0_conf
+
+# Reseting global.conf
+cp /usr/local/zevenet/app/checkglobalconf/global.conf.tpl ${CONF_DIR}/global.conf
+
+echo "Deleting cherokee configuration"
 SERVER=`grep "server!bind!1!interface =" /usr/local/zevenet/app/cherokee/etc/cherokee/cherokee.conf`
 sed -i "s/$SERVER/\#server\!bind\!1\!interface = \;/g" /usr/local/zevenet/app/cherokee/etc/cherokee/cherokee.conf
 PORT=`grep "server!bind!1!port =" /usr/local/zevenet/app/cherokee/etc/cherokee/cherokee.conf`
 sed -i "s/$PORT/server\!bind\!1\!port = 444/g" /usr/local/zevenet/app/cherokee/etc/cherokee/cherokee.conf
- 
-echo "Preparando para el primer arranque"
+
+START_MSG="#make your own script in your favorite language, it will be called
+#at the end of the procedure /etc/init.d/zenloadbalacer start
+#and replicated to the other node if zen cluster is running."
+
+STOP_MSG="#make your own script in your favorite language, it will be called
+#at the end of the procedure /etc/init.d/zenloadbalacer stop
+#and replicated to the other node if zen cluster is running."
+
+echo $START_MSG > zlb-start
+echo $STOP_MSG > zlb-stop
+
+
+echo "Cleaning apt"
+rm -fr /etc/apt/sources.list
+apt-get update
+apt-get clean
+
+
+echo "Preparing the firt boot"
 if [ ! -f /etc/firstzlbboot ]; then
         touch /etc/firstzlbboot
 fi
-echo "creating interface config file  if_${if_mgmt}_conf"
-echo "status=up" > /usr/local/zevenet/config/if_${if_mgmt}_conf
-echo "${if_mgmt};192.168.0.99;255.255.255.0;;" >> /usr/local/zevenet/config/if_${if_mgmt}_conf
-echo "Borrando el home de root"
+
+echo "Creating interface config file  if_${if_mgmt}_conf"x
+echo "status=up" > ${CONF_DIR}/if_${if_mgmt}_conf
+echo "${if_mgmt};192.168.0.99;255.255.255.0;;" >> ${CONF_DIR}/if_${if_mgmt}_conf
+
+echo "Deleting the root's home"
 rm -rf /root/.bash_history
 rm -rf /root/* {.bashrc}
